requirements.txt
numpy==1.19.5
keras==2.6.0
streamlit==0.88.0
tensorflow==2.6.0
Pillow==8.3.2
